A partial implementation for grabbing instance types for a given cloud.
The data is fetched fine, but something is screwed up again with the `select`'s: creating a `select` dropdown with the list of instance types results in an empty dropdown. Further, `change` event on the cloud list dropdown does not trigger and hence the list of types cannot be dynamically retrieved.