Remove the scheme from appliance access URL, suitably renaming it
We don't know if the appliance offers web access. It can
alternatively offer https. It also shortens the length of text to
display. And since it's no longer a URL, I renamed it to
instancePublicIP since that's what's really stored.

There is actually a bit of a problem here since we assume (and hence
store) the public IP locally. Yet, the IP can be remapped during the
life of an instance, and we wouldn't know about it. It might be
better to store the instance ID instead and fetch the public IP.