Add modal when deleting/archiving a deployment
We should probably watch (i.e., check) the spawned task and report status based on the outcome at the end of the action but this works for now.

We're also going to have to incorporate some changes to the layout and logic based on actions. For example, does deleting a deployment automatically archive it? Can we archive a deployment without deleting it?